Update to spirv-builder 0.9.0-git, add Cargo.lock
Without `Cargo.lock` the resolver is free to find recent dependencies
that aren't supported by the `rust-toolchain.toml` pins.

In particular, as of today (May 24th 2024), the combination of a pinned
Rust toolchain and an unpinned Cargo depdency tree resulted in this
error:
```
error: package `clap_lex v0.7.0` cannot be built because it requires rustc 1.74
or newer, while the currently active rustc version is 1.71.0-nightly
Either upgrade to rustc 1.74 or newer, or use
cargo update -p clap_lex@0.7.0 --precise ver
where `ver` is the latest version of `clap_lex` supporting rustc 1.71.0-nightly
```

There are other errors too if that one is fixed.

I've also updated to the latest `spirv-builder`, although I had to pin
to the latest HEAD, rather than the latest tag (v0.9.0). That tagged version
also fails to build with the "latest" `clap` (latest within the pin bounds).
I think if we'd had a `Cargo.lock` here, we could have frozen in time
the dependency tree of exact versions that worked with the pinned Rust toolchain
at the time of the last `spirv-builder` release.